seperate create_solana_tools
